Add ZepKit integration tests
