+++ tests/zepkit/test/README.md
@@ -1,0 +1,3 @@
+# ZepKit integration test
+
+Integration tests for ZepKit and an unpack command.
