+++ tests/zepkit/test/scripts/test.sh
